DIa 3 de la piscina

+++ Manipulate strings/ft_strlcpy.c
@@ -12,10 +12,10 @@
 
 #include "../libft.h"
 
-unsigned int	ft_strlcpy(char *dest, char *src, unsigned int size)
+size_t	ft_strlcpy(char *dest, char *src, size_t size)
 {
-	unsigned int	i;
-	unsigned int	x;
+	size_t	i;
+	size_t	x;
 
 	x = ft_strlen(src);
 	i = 0;
